+++ commonjs.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const OFF = 'off';
+
+module.exports = {
+  env: {
+    commonjs: true,
+  },
+  parserOptions: {
+    sourceType: 'script',
+  },
+  rules: {
+    // Import ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    'import/no-commonjs': OFF,
+  },
+};