+++ module.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const ERROR = 'error';
+
+module.exports = {
+  parserOptions: {
+    sourceType: 'module',
+  },
+  rules: {
+    // Unicorn ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    'unicorn/prefer-module': ERROR,
+  },
+};